requirements.txt
polygon
polygon-api-client
python-dotenv